pyproject.toml
[project]
name = "saomsg"
dynamic = ["version"]
authors = [
    { name = "T. E. Pickering", email = "te.pickering@gmail.com"}
]
license = {file = "licenses/LICENSE.rst"}
readme = "README.rst"
description = "Python interface to SAO MSG protocol"
requires-python = ">=3.11"
dependencies = [
    "astropy",
    "pyindi@git+https://github.com/MMTObservatory/pyINDI"
]

[project.optional-dependencies]
test = [
    "pytest-asyncio",
    "pytest-astropy",
]
docs = [
    "sphinx-astropy",
]

[project.urls]
Repository = "https://github.com/mmtobservatory/py-saomsg.git"
Documentation = "https://py-saomsg.readthedocs.io/"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages]
find = {}

[tool.setuptools_scm]
version_file = "saomsg/version.py"

[build-system]
requires = [
    "setuptools",
    "setuptools_scm",
]

build-backend = 'setuptools.build_meta'

[tool.pytest.ini_options]
minversion = 7.0
testpaths = [
    "saomsg/test",
]
astropy_header = true
doctest_plus = "enabled"
text_file_format = "rst"
addopts = [
    "--color=yes",
    "--doctest-rst",
]
xfail_strict = true
filterwarnings = [
    "error",
    "ignore:numpy\\.ufunc size changed:RuntimeWarning",
    "ignore:numpy\\.ndarray size changed:RuntimeWarning",
    # Python 3.12 warning from dateutil imported by matplotlib
    "ignore:.*utcfromtimestamp:DeprecationWarning",
]

[tool.coverage]

    [tool.coverage.run]
        omit = [
            "saomsg/_astropy_init*",
            "saomsg/conftest.py",
            "saomsg/tests/*",
            "saomsg/version*",
            "*/saomsg/_astropy_init*",
            "*/saomsg/conftest.py",
            "*/saomsg/tests/*",
            "*/saomsg/version*",
        ]

    [tool.coverage.report]
        exclude_lines = [
            # Have to re-enable the standard pragma
            "pragma: no cover",
            # Don't complain about packages we have installed
            "except ImportError",
            # Don't complain if tests don't hit defensive assertion code:
            "raise AssertionError",
            "raise NotImplementedError",
            # Don't complain about script hooks
            "'def main(.*):'",
            # Ignore branches that don't pertain to this version of Python
            "pragma: py{ignore_python_version}",
            # Don't complain about IPython completion helper
            "def _ipython_key_completions_",
        ]
